Delete `src/app/core/relations/relations.js`
